+++ webapp/channels/src/components/terms_of_service/terms_of_service.tsx
@@ -188,7 +188,7 @@ export default class TermsOfService extends React.PureComponent<TermsOfServicePr
                                 className='medium-center'
                                 data-testid='termsOfService'
                             >
-                                {messageHtmlToComponent(this.formattedText(this.state.customTermsOfServiceText), false, {mentions: false})}
+                                {messageHtmlToComponent(this.formattedText(this.state.customTermsOfServiceText), {mentions: false}, false)}
                             </div>
                         </div>
                         <div className='terms-of-service__footer medium-center'>
